# Topic/metal cache