--- a/include/mbgl/mtl/drawable.hpp
+++ b/include/mbgl/mtl/drawable.hpp
@@ -4,8 +4,6 @@
 #include <mbgl/gfx/draw_mode.hpp>
 #include <mbgl/mtl/upload_pass.hpp>
 #include <mbgl/programs/segment.hpp>
-#include <mbgl/mtl/mtl_fwd.hpp>
-#include <mbgl/mtl/render_pass.hpp>
 
 #include <memory>
 
@@ -59,6 +57,11 @@ public:
 
     void setShader(gfx::ShaderProgramBasePtr) override;
 
+    void setEnableStencil(bool) override;
+    void setEnableDepth(bool) override;
+    void setSubLayerIndex(int32_t) override;
+    void setDepthType(gfx::DepthMaskType) override;
+
 protected:
     // For testing only.
     Drawable(std::unique_ptr<Impl>);
@@ -76,12 +79,6 @@ protected:
 
     class Impl;
     const std::unique_ptr<Impl> impl;
-
-    gfx::AttributeBindingArray attributeBindings;
-
-    mutable MTLRenderPipelineStatePtr pipelineState;
-
-    mutable std::optional<gfx::RenderPassDescriptor> renderPassDescriptor;
 };
 
 } // namespace mtl