--- a/src/mbgl/mtl/offscreen_texture.cpp
+++ b/src/mbgl/mtl/offscreen_texture.cpp
@@ -38,6 +38,8 @@ public:
         commandBuffer->commit();
         commandBuffer->waitUntilCompleted();
         commandBuffer.reset();
+
+        cachedRenderPassDescriptor.reset();
     }
 
     PremultipliedImage readStillImage() {
@@ -59,12 +61,14 @@ public:
         return NS::TransferPtr(MTL::BlitPassDescriptor::alloc()->init());
     }
 
-    MTLRenderPassDescriptorPtr getRenderPassDescriptor() const override {
-        auto renderPassDescriptor = NS::TransferPtr(MTL::RenderPassDescriptor::alloc()->init());
-        if (auto* colorTarget = renderPassDescriptor->colorAttachments()->object(0)) {
-            colorTarget->setTexture(static_cast<Texture2D*>(texture.get())->getMetalTexture());
+    const MTLRenderPassDescriptorPtr& getRenderPassDescriptor() const override {
+        if (!cachedRenderPassDescriptor) {
+            cachedRenderPassDescriptor = NS::TransferPtr(MTL::RenderPassDescriptor::alloc()->init());
+            if (auto* colorTarget = cachedRenderPassDescriptor->colorAttachments()->object(0)) {
+                colorTarget->setTexture(static_cast<Texture2D*>(texture.get())->getMetalTexture());
+            }
         }
-        return renderPassDescriptor;
+        return cachedRenderPassDescriptor;
     }
 
 private:
@@ -73,6 +77,7 @@ private:
     const gfx::TextureChannelDataType type;
     gfx::Texture2DPtr texture;
     MTLCommandBufferPtr commandBuffer;
+    mutable MTLRenderPassDescriptorPtr cachedRenderPassDescriptor;
 };
 
 OffscreenTexture::OffscreenTexture(Context& context, const Size size_, const gfx::TextureChannelDataType type)