Avoid re-creating some Metal objects for every drawable on every frame.

Interestingly, the second commit caching the depth-stencil descriptor seems to make a negligible difference in the benchmark (within 0.1 ms), despite having a 99.3% hit rate.

Before:
```
| boston     | 24.8 ms | 40.3 fps |
| paris      | 29.0 ms | 34.5 fps |
| paris2     | 23.2 ms | 43.1 fps |
| alps       | 17.0 ms | 58.9 fps |
| us east    | 15.3 ms | 65.3 fps |
| greater la | 24.8 ms | 40.3 fps |
| sf         | 33.4 ms | 29.9 fps |
| oakland    | 24.1 ms | 41.5 fps |
| germany    | 23.0 ms | 43.5 fps |
Average frame time: 23.8 ms
Average FPS: 41.9
| boston     | 24.8 ms | 40.2 fps |
| paris      | 28.9 ms | 34.6 fps |
| paris2     | 24.5 ms | 40.8 fps |
| alps       | 16.5 ms | 60.5 fps |
| us east    | 17.2 ms | 58.1 fps |
| greater la | 26.0 ms | 38.4 fps |
| sf         | 34.3 ms | 29.1 fps |
| oakland    | 23.9 ms | 41.8 fps |
| germany    | 23.1 ms | 43.4 fps |
Average frame time: 24.4 ms
Average FPS: 41.0
| boston     | 24.8 ms | 40.3 fps |
| paris      | 27.6 ms | 36.2 fps |
| paris2     | 24.3 ms | 41.2 fps |
| alps       | 16.7 ms | 59.8 fps |
| us east    | 16.5 ms | 60.5 fps |
| greater la | 24.8 ms | 40.4 fps |
| sf         | 33.4 ms | 30.0 fps |
| oakland    | 23.9 ms | 41.9 fps |
| germany    | 23.0 ms | 43.5 fps |
Average frame time: 23.9 ms
Average FPS: 41.9
```

After:
```
| boston     | 24.8 ms | 40.3 fps |
| paris      | 27.7 ms | 36.1 fps |
| paris2     | 24.7 ms | 40.4 fps |
| alps       | 15.4 ms | 65.1 fps |
| us east    | 14.3 ms | 70.1 fps |
| greater la | 24.8 ms | 40.3 fps |
| sf         | 33.2 ms | 30.1 fps |
| oakland    | 23.5 ms | 42.5 fps |
| germany    | 23.0 ms | 43.4 fps |
Average frame time: 23.5 ms
Average FPS: 42.6
| boston     | 24.8 ms | 40.4 fps |
| paris      | 27.6 ms | 36.2 fps |
| paris2     | 24.8 ms | 40.4 fps |
| alps       | 14.8 ms | 67.6 fps |
| us east    | 14.2 ms | 70.3 fps |
| greater la | 24.8 ms | 40.3 fps |
| sf         | 33.2 ms | 30.1 fps |
| oakland    | 23.6 ms | 42.4 fps |
| germany    | 23.7 ms | 42.1 fps |
Average frame time: 23.5 ms
Average FPS: 42.5
| boston     | 24.8 ms | 40.3 fps |
| paris      | 27.6 ms | 36.3 fps |
| paris2     | 24.8 ms | 40.3 fps |
| alps       | 14.9 ms | 67.2 fps |
| us east    | 14.1 ms | 70.8 fps |
| greater la | 24.8 ms | 40.3 fps |
| sf         | 33.1 ms | 30.2 fps |
| oakland    | 23.6 ms | 42.3 fps |
| germany    | 23.4 ms | 42.8 fps |
Average frame time: 23.5 ms
Average FPS: 42.6
```

Mean frame time 24.033 => 23.5, 2.2% improvement